Introduce Folder structure and update some of the old files to use new api
