+++ api/configs/functions.js
@@ -1,0 +1,99 @@
+const databases = bot.api.databases.functions;
+const err = bot.err;
+
+/**
+ * Saves a config to the matching database. eg <modulename>_config
+ *
+ * @param mod_attributes {{default_config: {}, modulename: string}}
+ * @param guild_ID {string}
+ * @param config {{}}
+ *
+ * @throws {errors.js.BotError}, when a key in the provided config object is not in the default config
+ */
+function config_saveall(mod_attributes, guild_ID, config) {
+	if (!config) {
+		config = mod_attributes.default_config;
+	} else {
+		for (let key in config) {
+			if (!(key in mod_attributes.default_config))
+				throw new err.BotError("Falscher key.");
+		}
+		for (let key in mod_attributes.default_config) {
+			if (!(key in config)) config[key] = mod_attributes.default_config[key];
+		}
+	}
+	const dbs = mod_attributes.modulename + "_config";
+	databases.database_create_if_not_exists(dbs, ["guild", "key_value"]);
+	try {
+		let i = databases.lookup_key_value(dbs, "guild", guild_ID);
+		databases.database_row_change(dbs, i[0], "key_value", config);
+	} catch (e) {
+		databases.database_row_add(dbs, [guild_ID, config]);
+	}
+}
+
+/**
+ * Updates the config and saves it.
+ *
+ * @param mod_attributes {{default_config: {}, modulename: string}}
+ * @param guild_ID {string}
+ * @param key {string}
+ * @param value {*}
+ */
+export function config_update(mod_attributes, guild_ID, key, value) {
+	let config = config_load(mod_attributes, guild_ID);
+	if (!(key in config)) throw new err.BotError("Kein valider Key.");
+	config[key] = value;
+	config_saveall(mod_attributes, guild_ID, config);
+}
+
+/**
+ * Loads a config and returns the config map.
+ *
+ * @param mod_attributes {{default_config: {}, modulename: string}}
+ * @param guild_ID {string}
+ *
+ * @returns {{}}
+ */
+function config_load(mod_attributes, guild_ID) {
+	const dbs = mod_attributes.modulename + "_config";
+	let config = mod_attributes.default_config;
+	databases.database_create_if_not_exists(dbs, ["guild", "key_value"]);
+	try {
+		let i = databases.lookup_key_value(dbs, "guild", guild_ID);
+		if (i.length > 1) throw new err.BotError("Zu viele Einträge.");
+		config = databases.lookup_index(dbs, i[0], "key_value");
+		//TODO solve this line
+		// eslint-disable-next-line no-empty
+	} catch (e) {
+		// if it crashes, there is a database with this key.
+
+	}
+	return config;
+}
+
+/**
+ * Gets a config key.
+ *
+ * @param mod_attributes  {{default_config: {}, modulename: string}}
+ * @param guild_ID {string}
+ * @param key {string}
+ *
+ * @returns {*} the matching json object.
+ *
+ * @throws {errors.js.BotError}, if the key was not in the config
+ */
+export function config_get(mod_attributes, guild_ID, key) {
+	let config = config_load(mod_attributes, guild_ID);
+	if (!(key in config)) throw new err.BotError("Kein valider Key.");
+	return config[key];
+}
+
+export function config_toStr(mod_attributes, guild_ID) {
+	let config = config_load(mod_attributes, guild_ID);
+	let out = "";
+	for (let cfg_key in config) {
+		out += `${cfg_key} = ${JSON.stringify(config[cfg_key])}\n`;
+	}
+	return out.trim();
+}
