+++ api/configs/summary.js
@@ -1,0 +1,5 @@
+import * as functions from "./functions.js";
+
+module.exports = {
+	functions,
+};
