+++ api/constants.js
@@ -1,0 +1,2 @@
+export const { prefix } = require("../config.json");
+export const wait = require("util").promisify(setTimeout); //async wait
