+++ api/errors/BotError.js
@@ -1,0 +1,6 @@
+export default class BotError extends Error {
+	constructor(message) {
+		super(message);
+		this.name = this.constructor.name;
+	}
+}
