+++ api/errors/Command.js
@@ -1,0 +1,7 @@
+import BotError from "./BotError";
+
+export default class Command extends BotError {
+	constructor(message) {
+		super(message);
+	}
+}
