+++ api/errors/CommandNameNotFound.js
@@ -1,0 +1,7 @@
+import Command from "./Command";
+
+export default class CommandNameNotFound extends Command {
+	constructor(cmdname, modulename) {
+		super(`Konnte den Befehl ${cmdname} für das ${modulename} nicht finden`);
+	}
+}
