+++ api/errors/Dublication.js
@@ -1,0 +1,7 @@
+import BotError from "./BotError";
+
+export default class Dublication extends BotError {
+	constructor(variable) {
+		super(`${variable} existiert bereits`);
+	}
+}
