+++ api/errors/InvalidData.js
@@ -1,0 +1,7 @@
+import BotError from "./BotError";
+
+export default class InvalidData extends BotError {
+	constructor() {
+		super("Die Daten waren ungültig");
+	}
+}
