+++ api/errors/Undefined.js
@@ -1,0 +1,7 @@
+import BotError from "./BotError";
+
+export default class Undefined extends BotError {
+	constructor(variable) {
+		super(`${variable} war undefiniert`);
+	}
+}
