+++ api/errors/Unexisting.js
@@ -1,0 +1,7 @@
+import BotError from "./BotError";
+
+export default class Unexisting extends BotError {
+	constructor(variable) {
+		super(`${variable} existiert nicht`);
+	}
+}
