+++ api/errors/summary.js
@@ -1,0 +1,29 @@
+import BotError from "./BotError";
+import Command from "./Command";
+import CommandNameNotFound from "./CommandNameNotFound";
+import Dublication from "./Dublication";
+import Find from "./Find";
+import InvalidData from "./InvalidData";
+import ParameterArguments from "./ParameterArguments";
+import ParameterDependency from "./ParameterDependency";
+import ParameterRequired from "./ParameterRequired";
+import Range from "./Range";
+import Type from "./Type";
+import Undefined from "./Undefined";
+import Unexisting from "./Unexisting";
+
+module.exports = {
+	BotError,
+	Undefined,
+	Type,
+	InvalidData,
+	Find,
+	Range,
+	Unexisting,
+	Dublication,
+	Command,
+	CommandNameNotFound,
+	ParameterArguments,
+	ParameterDependency,
+	ParameterRequired,
+};
