+++ modules/help/attributes.js
@@ -1,29 +1,33 @@
+const Command = bot.api.commands.Command;
+const Parameter = bot.api.commands.Parameter;
+const prefix = bot.api.constants.prefix;
+
 const attributes = {
 	modulename: "help",
 	description:
 		"Das Help-Modul. Hier kannst Du vieles über Commands und andere Module erfahren.",
 	commands: [
-		new bot.api.Command(
+		new Command(
 			"modulehelp",
 			"Zeigt die Hilfeseite von dem angegebenen Modul",
 			[
-				new bot.api.Parameter(
+				new Parameter(
 					"-name",
 					"required",
 					[],
 					"Der Name des Moduls.",
-					(nr) => nr == 1,
+					(nr) => nr === 1,
 					["help"],
 					true
 				),
 			],
-			[`${bot.api.prefix}help modulehelp -name chmgr`]
+			[`${prefix}help modulehelp -name chmgr`]
 		),
-		new bot.api.Command(
+		new Command(
 			"listmodules",
 			"Zeigt alle verfügbaren Module an",
 			[],
-			[`${bot.api.prefix}help listmodules`]
+			[`${prefix}help listmodules`]
 		),
 	],
 };
